prepare to merge with tural
