+++ package.json
@@ -15,6 +15,7 @@
     "format": "prettier --write ."
   },
   "dependencies": {
+    "@azure/storage-blob": "^12.25.0",
     "@date-io/dayjs": "^3.0.0",
     "@emotion/react": "^11.11.0",
     "@emotion/styled": "^11.11.0",
@@ -39,6 +40,7 @@
     "react": "^18.3.1",
     "react-datepicker": "^7.3.0",
     "react-dom": "^18.3.1",
+    "react-dropzone": "^14.2.3",
     "rrule": "^2.7.2"
   },
   "devDependencies": {
