+++ src/app/components/UI/PostFilter.js
@@ -2,7 +2,7 @@
 
 'use client';
 
-import React, { useContext } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { categoryColors } from '@/utils/categoryColors';
